Complete proper deleting of users.

+++ backend/firey-leaf-events/src/main/java/nl/benzelinsky/fireyleafevents/services/UserService.java
@@ -4,10 +4,7 @@ import nl.benzelinsky.fireyleafevents.dtos.PatchUserInputDto;
 import nl.benzelinsky.fireyleafevents.dtos.ShortUserOutputDto;
 import nl.benzelinsky.fireyleafevents.dtos.UserInputDto;
 import nl.benzelinsky.fireyleafevents.dtos.UserOutputDto;
-import nl.benzelinsky.fireyleafevents.exceptions.RoleNotFoundException;
-import nl.benzelinsky.fireyleafevents.exceptions.UserAlreadyExistsException;
-import nl.benzelinsky.fireyleafevents.exceptions.UsernameNotFoundException;
-import nl.benzelinsky.fireyleafevents.exceptions.UsernameUnavailableException;
+import nl.benzelinsky.fireyleafevents.exceptions.*;
 import nl.benzelinsky.fireyleafevents.mappers.UserMapper;
 import nl.benzelinsky.fireyleafevents.models.Role;
 import nl.benzelinsky.fireyleafevents.models.User;
@@ -17,9 +14,7 @@ import nl.benzelinsky.fireyleafevents.utils.RandomStringGenerator;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Set;
+import java.util.*;
 
 @Service
 public class UserService {
@@ -147,7 +142,19 @@ public class UserService {
         User toDelete = this.userRepository.findById(username)
                 .orElseThrow(() ->
                         new UsernameNotFoundException(username));
-        this.userRepository.deleteById(username);
+        if (!toDelete.getHostedEvents().isEmpty()) {
+            List<String> events = new ArrayList<>();
+            Map<Long, String> eventIds = new HashMap<>();
+            toDelete.getHostedEvents()
+                    .forEach(event -> {
+                        events.add(event.getName());
+                        eventIds.put(event.getId(), event.getName());
+                    });
+            throw new HasActiveEventsException(username, eventIds);
+        }
+        else {
+            this.userRepository.deleteById(username);
+        }
         return "User " + toDelete.getUsername() + " has been deleted.";
     }
 }
